Push Evidence down to System.Security.AccessControl
This can avoid the dependency
  System.Security.Cryptography.Xml > System.Security.Permissions
